move config & errors to combined file
