+++ src/client/mod.rs
@@ -1,0 +1,37 @@
+// MIDEN CLIENT
+// ================================================================================================
+
+use crate::{config::ClientConfig, errors::ClientError, store::Store};
+
+pub mod accounts;
+pub mod notes;
+
+/// A light client for connecting to the Miden rollup network.
+///
+/// Miden client is responsible for managing a set of accounts. Specifically, the client:
+/// - Keeps track of the current and historical states of a set of accounts and related objects
+///   such as notes and transactions.
+/// - Connects to one or more Miden nodes to periodically sync with the current state of the
+///   network.
+/// - Executes, proves, and submits transactions to the network as directed by the user.
+pub struct Client {
+    /// Local database containing information about the accounts managed by this client.
+    store: Store,
+    // TODO
+    // node: connection to Miden node
+}
+
+impl Client {
+    // CONSTRUCTOR
+    // --------------------------------------------------------------------------------------------
+
+    /// Returns a new instance of [Client] instantiated with the specified configuration options.
+    ///
+    /// # Errors
+    /// Returns an error if the client could not be instantiated.
+    pub fn new(config: ClientConfig) -> Result<Self, ClientError> {
+        Ok(Self {
+            store: Store::new((&config).into())?,
+        })
+    }
+}
